requirements.txt
numpy==1.23.5
pandas==2.0.3
Pillow==10.2.0